Updated AppConfigSheet
copywrite, portfolio links, and other things added

Additionally, a disclaimer has been included that this is not to be assessed as part of Rio's project. While it is helpful, all functions should still be possible on device, and in his own work. I made this because I was bored, and Rio was making an api anyways